Initial 64-bit support for ndisasm.  Still a work in progress.
64-bit support for ndisasm.  This is very much an initial attempt, and
there are guaranteed to be bugs in the code.  However, some *very*
preliminary testing seems to indicate it's not completely off-base.

+++ nasm.h
@@ -223,7 +223,7 @@ enum {                          /* token types, other than chars */
 
 typedef struct {
     int32_t segment;
-    int32_t offset;
+    int64_t offset;
     int known;
 } loc_t;
